Changed calculation of coalescence times
The scaling for "partial coalescent" made no sense. The result was that
if the number of targets was not equal to the number of neighbors, the
smaller patch coalesced more slowly than the larger one.

Now I halve the coalescence time within the patches, which always gives
a symmetrical tree (see chunk 10a).